Fix autoindex segment num
