+++ internal/querynodev2/delegator/delegator_test.go
@@ -160,7 +160,7 @@ func (s *DelegatorSuite) SetupTest() {
 		NewMsgStreamFunc: func(_ context.Context) (msgstream.MsgStream, error) {
 			return s.mq, nil
 		},
-	}, 10000)
+	}, 10000, nil)
 	s.Require().NoError(err)
 }
