+++ internal/querynodev2/handlers.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"strconv"
 
-	"github.com/golang/protobuf/proto"
 	"github.com/samber/lo"
 	"go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
@@ -29,13 +28,11 @@ import (
 	"github.com/milvus-io/milvus-proto/go-api/v2/commonpb"
 	"github.com/milvus-io/milvus/internal/proto/datapb"
 	"github.com/milvus-io/milvus/internal/proto/internalpb"
-	"github.com/milvus-io/milvus/internal/proto/planpb"
 	"github.com/milvus-io/milvus/internal/proto/querypb"
 	"github.com/milvus-io/milvus/internal/querynodev2/delegator"
 	"github.com/milvus-io/milvus/internal/querynodev2/segments"
 	"github.com/milvus-io/milvus/internal/querynodev2/tasks"
 	"github.com/milvus-io/milvus/internal/util/streamrpc"
-	"github.com/milvus-io/milvus/pkg/common"
 	"github.com/milvus-io/milvus/pkg/log"
 	"github.com/milvus-io/milvus/pkg/metrics"
 	"github.com/milvus-io/milvus/pkg/util/funcutil"
@@ -332,72 +329,6 @@ func (node *QueryNode) queryStreamSegments(ctx context.Context, req *querypb.Que
 	return nil
 }
 
-func (node *QueryNode) optimizeSearchParams(ctx context.Context, req *querypb.SearchRequest, deleg delegator.ShardDelegator) (*querypb.SearchRequest, error) {
-	// no hook applied, just return
-	if node.queryHook == nil {
-		return req, nil
-	}
-
-	log := log.Ctx(ctx).With(zap.Int64("collection", req.GetReq().GetCollectionID()))
-
-	serializedPlan := req.GetReq().GetSerializedExprPlan()
-	// plan not found
-	if serializedPlan == nil {
-		log.Warn("serialized plan not found")
-		return req, merr.WrapErrParameterInvalid("serialized search plan", "nil")
-	}
-
-	channelNum := req.GetTotalChannelNum()
-	// not set, change to conservative channel num 1
-	if channelNum <= 0 {
-		channelNum = 1
-	}
-
-	plan := planpb.PlanNode{}
-	err := proto.Unmarshal(serializedPlan, &plan)
-	if err != nil {
-		log.Warn("failed to unmarshal plan", zap.Error(err))
-		return nil, merr.WrapErrParameterInvalid("valid serialized search plan", "no unmarshalable one", err.Error())
-	}
-
-	switch plan.GetNode().(type) {
-	case *planpb.PlanNode_VectorAnns:
-		// ignore growing ones for now since they will always be brute force
-		sealed, _ := deleg.GetSegmentInfo(true)
-		sealedNum := lo.Reduce(sealed, func(sum int, item delegator.SnapshotItem, _ int) int {
-			return sum + len(item.Segments)
-		}, 0)
-		// use shardNum * segments num in shard to estimate total segment number
-		estSegmentNum := sealedNum * int(channelNum)
-		withFilter := (plan.GetVectorAnns().GetPredicates() != nil)
-		queryInfo := plan.GetVectorAnns().GetQueryInfo()
-		params := map[string]any{
-			common.TopKKey:        queryInfo.GetTopk(),
-			common.SearchParamKey: queryInfo.GetSearchParams(),
-			common.SegmentNumKey:  estSegmentNum,
-			common.WithFilterKey:  withFilter,
-			common.CollectionKey:  req.GetReq().GetCollectionID(),
-		}
-		err := node.queryHook.Run(params)
-		if err != nil {
-			log.Warn("failed to execute queryHook", zap.Error(err))
-			return nil, merr.WrapErrServiceUnavailable(err.Error(), "queryHook execution failed")
-		}
-		queryInfo.Topk = params[common.TopKKey].(int64)
-		queryInfo.SearchParams = params[common.SearchParamKey].(string)
-		serializedExprPlan, err := proto.Marshal(&plan)
-		if err != nil {
-			log.Warn("failed to marshal optimized plan", zap.Error(err))
-			return nil, merr.WrapErrParameterInvalid("marshalable search plan", "plan with marshal error", err.Error())
-		}
-		req.Req.SerializedExprPlan = serializedExprPlan
-		log.Debug("optimized search params done", zap.Any("queryInfo", queryInfo))
-	default:
-		log.Warn("not supported node type", zap.String("nodeType", fmt.Sprintf("%T", plan.GetNode())))
-	}
-	return req, nil
-}
-
 func (node *QueryNode) searchChannel(ctx context.Context, req *querypb.SearchRequest, channel string) (*internalpb.SearchResults, error) {
 	log := log.Ctx(ctx).With(
 		zap.Int64("msgID", req.GetReq().GetBase().GetMsgID()),
@@ -434,11 +365,6 @@ func (node *QueryNode) searchChannel(ctx context.Context, req *querypb.SearchReq
 	if !ok {
 		err := merr.WrapErrChannelNotFound(channel)
 		log.Warn("Query failed, failed to get shard delegator for search", zap.Error(err))
-		return nil, err
-	}
-	req, err = node.optimizeSearchParams(ctx, req, sd)
-	if err != nil {
-		log.Warn("failed to optimize search params", zap.Error(err))
 		return nil, err
 	}
 	// do search
